Rename client interfaces.
